backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
qdrant-client==1.8.1
google-generativeai==0.4.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
